test: refactor expand/integrity tests to parser_test.go
Moved the tests for `expand` and `integrity` header fields into `parser_test.go`
to consolidate test files. Used `//go:embed` to load the test data from `testdata/`
instead of using inline multiline strings.

Addresses PR feedback.
